Break up center to reduce coupling
No breaking changes. God objects are not good.
