+++ ArimBans3/src/main/java/space/arim/bans/ArimBans.java
@@ -19,302 +19,113 @@
 package space.arim.bans;
 
 import java.io.File;
-import java.io.IOException;
 import java.sql.ResultSet;
-import java.util.Set;
-import java.util.UUID;
-import java.util.logging.FileHandler;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 import space.arim.bans.api.ArimBansLibrary;
-import space.arim.bans.api.AsyncExecutor;
-import space.arim.bans.api.CommandType;
-import space.arim.bans.api.Punishment;
 import space.arim.bans.api.PunishmentPlugin;
-import space.arim.bans.api.PunishmentType;
-import space.arim.bans.api.Subject;
 import space.arim.bans.api.UUIDResolver;
-import space.arim.bans.api.exception.ConflictingPunishmentException;
 import space.arim.bans.env.Environment;
 import space.arim.bans.internal.Configurable;
-import space.arim.bans.internal.async.AsyncMaster;
-import space.arim.bans.internal.async.AsyncWrapper;
-import space.arim.bans.internal.async.Async;
-import space.arim.bans.internal.backend.punishment.Punishments;
 import space.arim.bans.internal.backend.punishment.PunishmentsMaster;
-import space.arim.bans.internal.backend.resolver.Resolver;
 import space.arim.bans.internal.backend.resolver.ResolverMaster;
 import space.arim.bans.internal.backend.subjects.SubjectsMaster;
-import space.arim.bans.internal.backend.subjects.Subjects;
-import space.arim.bans.internal.config.Config;
 import space.arim.bans.internal.config.ConfigMaster;
-import space.arim.bans.internal.frontend.commands.Commands;
 import space.arim.bans.internal.frontend.commands.CommandsMaster;
-import space.arim.bans.internal.frontend.format.Formats;
 import space.arim.bans.internal.frontend.format.FormatsMaster;
-import space.arim.bans.internal.sql.SqlQuery;
 import space.arim.bans.internal.sql.SqlMaster;
-import space.arim.bans.internal.sql.Sql;
+import space.arim.bans.internal.sql.SqlQuery;
 
-import space.arim.registry.RegistryPriority;
 import space.arim.registry.UniversalRegistry;
+import space.arim.registry.RegistryPriority;
 
-public class ArimBans implements Configurable, ArimBansLibrary {
+public interface ArimBans extends Configurable, ArimBansLibrary {
+
+	File dataFolder();
 	
-	private final File folder;
-	private Logger logger;
-	private final Environment environment;
-	private final ConfigMaster config;
-	private final SqlMaster sql;
-	private final PunishmentsMaster punishments;
-	private final SubjectsMaster subjects;
-	private final ResolverMaster resolver;
-	private final CommandsMaster commands;
-	private final FormatsMaster formats;
-	private final AsyncMaster async;
+	Environment environment();
 	
-	public ArimBans(File dataFolder, Environment environment) {
-		this.folder = dataFolder;
-		this.environment = environment;
-		if (dataFolder.mkdirs()) {
-			logger = Logger.getLogger(getName());
-			logger.setParent(environment.logger());
-			logger.setUseParentHandlers(false);
-			try {
-				logger.addHandler(new FileHandler(dataFolder + File.separator + "information.log"));
-			} catch (IOException ex) {
-				environment.logger().warning("ArimBans: **Severe Error**\nLogger initialisation in " + dataFolder.getPath() + File.separator + "information.log failed!");
-			}
-		} else {
-			environment().logger().warning("ArimBans: **Severe Error**\nDirectory creation of " + dataFolder.getPath() + " failed!");
-		}
-		config = new Config(this);
-		sql = new Sql(this);
-		punishments = new Punishments(this);
-		subjects = new Subjects(this);
-		resolver = new Resolver(this);
-		commands = new Commands(this);
-		formats = new Formats(this);
-		if (UniversalRegistry.isProvidedFor(AsyncExecutor.class)) {
-			async = new AsyncWrapper(UniversalRegistry.getRegistration(AsyncExecutor.class));
-		} else {
-			async = new Async(this);
-			UniversalRegistry.register(AsyncExecutor.class, (AsyncExecutor) async); 
-		}
-		refresh(true);
-		loadData();
-		UniversalRegistry.register(PunishmentPlugin.class, this);
-		UniversalRegistry.register(UUIDResolver.class, resolver);
-	}
+	ConfigMaster config();
 	
-	private void loadData() {
+	SqlMaster sql();
+	
+	PunishmentsMaster punishments();
+	
+	SubjectsMaster subjects();
+	
+	ResolverMaster resolver();
+	
+	CommandsMaster commands();
+	
+	FormatsMaster formats();
+	
+	default void loadData() {
 		sql().executeQuery(new SqlQuery(SqlQuery.Query.CREATE_TABLE_CACHE.eval(sql().settings())), new SqlQuery(SqlQuery.Query.CREATE_TABLE_ACTIVE.eval(sql().settings())), new SqlQuery(SqlQuery.Query.CREATE_TABLE_HISTORY.eval(sql().settings())));
 		ResultSet[] data = sql().selectQuery(new SqlQuery(SqlQuery.Query.SELECT_ALL_CACHED.eval(sql().settings())), new SqlQuery(SqlQuery.Query.SELECT_ALL_ACTIVE.eval(sql().settings())), new SqlQuery(SqlQuery.Query.SELECT_ALL_HISTORY.eval(sql().settings())));
 		resolver().loadAll(data[0]);
 		punishments().loadActive(data[1]);
 		punishments().loadHistory(data[2]);
 	}
-
-	public File dataFolder() {
-		return this.folder;
+	
+	default void register() {
+		UniversalRegistry.register(PunishmentPlugin.class, this);
+		UniversalRegistry.register(UUIDResolver.class, resolver());
 	}
 	
-	public Environment environment() {
-		return this.environment;
-	}
+	void log(String message);
 	
-	public ConfigMaster config() {
-		return config;
-	}
+	void logError(Exception ex);
 	
-	public SqlMaster sql() {
-		return sql;
-	}
-	
-	public PunishmentsMaster punishments() {
-		return punishments;
-	}
-
-	public SubjectsMaster subjects() {
-		return subjects;
-	}
-
-	public ResolverMaster resolver() {
-		return resolver;
-	}
-
-	public CommandsMaster commands() {
-		return commands;
-	}
-
-	public FormatsMaster formats() {
-		return formats;
-	}
-	
-	public void log(String message) {
-		if (logger != null) {
-			logger.info(message);
-		} else {
-			environment().logger().info(message);
-		}
-	}
-	
-	public void logError(Exception ex) {
-		if (logger != null) {
-			environment().logger().warning("Encountered and caught an error: " + ex.getLocalizedMessage() + " \nPlease check the plugin's log for more information. Please create a Github issue to address this.");
-			logger.log(Level.WARNING, "Encountered an error:", ex);
-		} else {
-			environment().logger().warning("Encountered and caught an error. \nNote that this plugin's log is inoperative, so the error will be printed to console. Please create a Github issue to address this.");
-			ex.printStackTrace();
-		}
+	@Override
+	default void refreshConfig(boolean first) {
+		config().refreshConfig(first);
+		sql().refreshConfig(first);
+		punishments().refreshConfig(first);
+		subjects().refreshConfig(first);
+		resolver().refreshConfig(first);
+		commands().refreshConfig(first);
+		formats().refreshConfig(first);
 	}
 	
 	@Override
-	public void refreshConfig(boolean first) {
-		config.refreshConfig(first);
-		sql.refreshConfig(first);
-		punishments.refreshConfig(first);
-		subjects.refreshConfig(first);
-		resolver.refreshConfig(first);
-		commands.refreshConfig(first);
-		formats.refreshConfig(first);
-		async.refreshConfig(first);
+	default void refreshMessages(boolean first) {
+		config().refreshMessages(first);
+		sql().refreshMessages(first);
+		punishments().refreshMessages(first);
+		subjects().refreshMessages(first);
+		resolver().refreshMessages(first);
+		commands().refreshMessages(first);
+		formats().refreshMessages(first);
 	}
 	
 	@Override
-	public void refreshMessages(boolean first) {
-		config.refreshMessages(first);
-		sql.refreshMessages(first);
-		punishments.refreshMessages(first);
-		subjects.refreshMessages(first);
-		resolver.refreshMessages(first);
-		commands.refreshMessages(first);
-		formats.refreshMessages(first);
-		async.refreshMessages(first);
+	default void close() {
+		config().close();
+		sql().close();
+		punishments().close();
+		subjects().close();
+		commands().close();
+		resolver().close();
+		formats().close();
 	}
 	
 	@Override
-	public void close() {
-		config.close();
-		async.close();
-		sql.close();
-		punishments.close();
-		subjects.close();
-		commands.close();
-		resolver.close();
-		formats.close();
-		async.close();
+	default String getName() {
+		return environment().getName();
 	}
 	
 	@Override
-	public String getName() {
-		return environment.getName();
+	default String getAuthor() {
+		return environment().getAuthor();
 	}
 	
 	@Override
-	public String getAuthor() {
-		return environment.getAuthor();
+	default String getVersion() {
+		return environment().getVersion();
 	}
 	
 	@Override
-	public String getVersion() {
-		return environment.getVersion();
-	}
-	
-	@Override
-	public byte getPriority() {
+	default byte getPriority() {
 		return RegistryPriority.LOWER;
 	}
 	
-	@Override
-	public boolean isBanned(Subject subject) {
-		return punishments().hasPunishment(subject, PunishmentType.BAN);
-	}
-	
-	@Override
-	public boolean isMuted(Subject subject) {
-		return punishments().hasPunishment(subject, PunishmentType.MUTE);
-	}
-	
-	@Override
-	public Set<Punishment> getBanList() {
-		return punishments().getAllPunishments(PunishmentType.BAN);
-	}
-	
-	@Override
-	public Set<Punishment> getMuteList() {
-		return punishments().getAllPunishments(PunishmentType.MUTE);
-	}
-	
-	@Override
-	public Set<Punishment> getWarns(Subject subject) {
-		return punishments().getPunishments(subject, PunishmentType.WARN);
-	}
-	
-	@Override
-	public Set<Punishment> getKicks(Subject subject) {
-		return punishments().getPunishments(subject, PunishmentType.KICK);
-	}
-	
-	@Override
-	public Set<Punishment> getHistory(Subject subject) {
-		return punishments().getHistory(subject);
-	}
-	
-	@Override
-	public void addPunishments(Punishment...punishments) throws ConflictingPunishmentException {
-		punishments().addPunishments(punishments);
-	}
-	
-	@Override
-	public Subject fromUUID(UUID subject) {
-		return Subject.fromUUID(subject);
-	}
-	
-	@Override
-	public Subject fromIpAddress(String address) throws IllegalArgumentException {
-		return Subject.fromIP(address);
-	}
-	
-	@Override
-	public Subject parseSubject(String input) throws IllegalArgumentException {
-		return subjects().parseSubject(input);
-	}
-	
-	@Override
-	public void simulateCommand(Subject subject, CommandType command, String[] args) {
-		commands().execute(subject, command, args);
-	}
-	
-	@Override
-	public void async(Runnable command) {
-		async.execute(command);
-	}
-	
-	@Override
-	public Logger getLogger() {
-		return logger;
-	}
-	
-	@Override
-	public void reload() {
-		refresh(true);
-	}
-	
-	@Override
-	public void reloadConfig() {
-		refreshConfig(true);
-	}
-	
-	@Override
-	public void reloadMessages() {
-		refreshMessages(true);
-	}
-	
-	@Override
-	public void sendMessage(Subject subject, String message) {
-		subjects.sendMessage(subject, message);
-	}
-
 }
