+++ ArimBans3/src/main/java/space/arim/bans/internal/async/AsyncMaster.java
@@ -18,17 +18,15 @@
  */
 package space.arim.bans.internal.async;
 
-import space.arim.bans.internal.Component;
-
-public interface AsyncMaster extends Component {
-	
-	@Override
-	default Class<?> getType() {
-		return AsyncMaster.class;
-	}
+public interface AsyncMaster extends AutoCloseable {
 	
 	void execute(Runnable command);
 	
 	boolean isClosed();
+	
+	@Override
+	default void close() {
+		
+	}
 	
 }
